move the file overwrite out of the begin..rescue.end block to the end of the method where the original is only overwritten if there were changes made to the temp copy

+++ libraries/provider_add_to_list.rb
@@ -60,14 +60,16 @@ class Chef
             end
             temp_file.puts line
           end
-
-          overwrite_original(f, temp_file) if ::File.cmp(f, temp_file)
         ensure
           temp_file.close unless temp_file.nil?
           f.close unless f.nil?
           temp_file.unlink
         end
-        new_resource.updated_by_last_action(true)
+
+        unless ::File.cmp(f, temp_file)
+          overwrite_original(f, temp_file)
+          new_resource.updated_by_last_action(true)
+        end
       end # def action_edit
 
       def nothing
